refactor: OOP 적용, 코드 구조 개선
