--- backend/controllers/README.md
@@ -1,1 +1,0 @@
-# 이 폴더는 업로드된 파일을 저장하기 위한 공간입니다.
